Avoid calling Tcl_AppendResult on a NULL interp.
The channel driver documentation points out that the interp parameter in
the close and get and set option functions may be NULL and we must avoid
using this in that case.

Also - newer versions of ftd2xx.h do not define FT_BITS_[56] so the
use of these values is denied when compiled against such a version.

Reported-by: Dean TerHaar <DTerHaar@xrite.com>
Signed-off-by: Pat Thoyts <patthoyts@users.sourceforge.net>